This version is VERY buggy, don't use unless you are brave. :)
This version adds initial support for multiple channels, a prerequisite to implementing the twitter search api. Currently the multiple channel suppport is only useful as a way to groups of the people you follow.  
Your entire timeline will always be in #twitter, but you can now /join multiple channels and /invite your friends to them, and their messsages will show up in that channel as well.
This may be useful if you wanted a subset of your friends messages to show up in another channel to make them easier to find.

There are still some issues with the internal state, and +v modes not properly being represented in the other channels.  Subsequent commits will fix these issues before this is ready for prime-time